requirements.txt
pydantic>=2.0
pyyaml>=6.0
hvac>=2.0
cryptography>=41.0
httpx>=0.25
